# swagger.yaml
openapi: 3.0.0
info:
  title: web scraper api
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:5000/api/v1
  variables: {}
paths:
...
  /searches:
    post:
      tags:
      - search
      summary: Create Search
      operationId: CreateSearch
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateSearchRequest'
              - example:
                  name: ktm 690 unweit von Berlin
                  link: https://www.ebay-kleinanzeigen.de/s-motorraeder-roller/berlin/preis::4000/ktm-690/k0c305l3331r200
            example:
              name: ktm 690 unweit von Berlin
              link: https://www.ebay-kleinanzeigen.de/s-motorraeder-roller/berlin/preis::4000/ktm-690/k0c305l3331r200
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - search
      summary: Get all searches
      operationId: Getallsearches
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    CreateSearchRequest:
      title: CreateSearchRequest
      required:
      - name
      - link
      type: object
      properties:
        name:
          type: string
        link:
          type: string
      example:
        name: ktm 690 unweit von Berlin
        link: https://www.ebay-kleinanzeigen.de/s-motorraeder-roller/berlin/preis::4000/ktm-690/k0c305l3331r200
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...